40-oo/animationen/conway/requirements.txt
numpy~=1.23.2
graphics.py